pyproject.toml
[tool.poetry]
name = "tooltorial"
version = "0.1.0"
description = "A tool-torial"
authors = ["Ben Eisner <ben.a.eisner@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.9.0"
torchvision = "^0.10.0"
typer = "^0.3.2"
pytorch-lightning = "^1.3.7"
wandb = "^0.10.33"
hydra-core = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
isort = "^5.9.1"
pytest = "^6.2.4"
mypy = "^0.910"
pylint = "^2.9.1"
autoflake = "^1.4"
rope = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
